service/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
mysqlclient==2.1.0
mysql-connector-python==8.0.28
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0